TST: Merge tests for constitutive laws
