+++ tests/models/test_constitutive_laws.py
@@ -1,6 +1,6 @@
 """Tests for constitutive laws in the models.
 
-The module contains a setup for testing composition and evaluation (in the Ad sense) of
+The first function is for testing composition and evaluation (in the Ad sense) of
 constitutive laws for models. The setup is applied to the constitutive laws in the
 models subpackage.
 
@@ -8,8 +8,10 @@ In addition to checking that evaluation is possible, the tests also check that t
 compound classes contain the specified constitutive laws, thus revealing mistakes if the
 classes are refactored.
 
-The module does not test that the constitutive laws evaluates to reasonable numbers. For
-this see test_evaluate_constitutive_laws.py.
+The first test does not test that the constitutive laws evaluates to reasonable numbers.
+That is done in the second and third tests. TODO: If those are extended to cover more
+constitutive laws, the first test might be removed.
+
 
 """
 from __future__ import annotations
@@ -18,7 +20,10 @@ import pytest
 
 import porepy as pp
 
-from . import setup_utils
+import porepy.models.constitutive_laws as c_l
+
+
+from porepy.applications.test_utils import models
 
 
 @pytest.mark.parametrize(
@@ -102,7 +107,7 @@ def test_parse_constitutive_laws(
 
     Parameters:
         model_type: Type of model to test. Currently supported are "mass_balance" and
-            "momentum_balance". To add a new model, add a new class in setup_utils.py.
+            "momentum_balance". To add a new model, add a new class in models.py.
         method_name (str): Name of the method to test.
         only_codimension (list of int): Indices of the domains for which the method
             should be called. Some methods are only defined for a subset of domains. If
@@ -132,11 +137,11 @@ def test_parse_constitutive_laws(
         return
 
     # Set up an object of the prescribed model
-    setup = setup_utils.model(model_type, domain_dim, num_fracs=num_fracs)
+    setup = models.model(model_type, domain_dim, num_fracs=num_fracs)
     # Fetch the relevant method of this model and extract the domains for which it is
     # defined.
     method = getattr(setup, method_name)
-    domains = setup_utils.subdomains_or_interfaces_from_method_name(
+    domains = models.subdomains_or_interfaces_from_method_name(
         setup.mdg, method, dimensions_to_assemble
     )
 
@@ -154,3 +159,254 @@ def test_parse_constitutive_laws(
     # with the constitutive law, or it could signify that something has changed in the
     # Ad machinery which makes the evaluation of the operator fail.
     op.evaluate(setup.equation_system)
+
+
+@pytest.mark.parametrize(
+    "model, method_name, expected, dimension",
+    [
+        (models.Poromechanics, "fluid_density", 1000 * np.exp(4e-10 * 2), None),
+        (
+            models.Thermoporomechanics,
+            "fluid_density",
+            # \rho = \rho_0 \exp(compressibility p - thermal_expansion T)
+            1000 * np.exp(4e-10 * 2 - 2.1e-4 * 3),
+            None,
+        ),
+        (
+            models.MassAndEnergyBalance,
+            "thermal_conductivity",
+            # Porosity weighted average of the solid and fluid thermal conductivities
+            (1 - 7e-3) * 2.5 + 7e-3 * 0.6,
+            None,
+        ),
+        (
+            models.MassAndEnergyBalance,
+            "solid_enthalpy",
+            # c_p T
+            790 * 3,
+            None,
+        ),
+        (
+            models.MassAndEnergyBalance,
+            "fluid_enthalpy",
+            # c_p T
+            4180 * 3,
+            None,
+        ),
+        (
+            models.Poromechanics,
+            "matrix_porosity",
+            # phi_0 + (alpha - phi_ref) * (1 - alpha) / bulk p. Only pressure, not
+            # div(u), is included in this test.
+            7e-3 + (0.8 - 7e-3) * (1 - 0.8) / (11.11 * pp.GIGA) * 2,
+            2,  # Matrix porosity is only defined in Nd
+        ),
+        (
+            models.Thermoporomechanics,
+            "matrix_porosity",
+            # phi_0 + (alpha - phi_ref) * (1 - alpha) / bulk * p
+            # - (alpha - phi_0) * thermal expansion * T
+            #  Only pressure and temperature, not div(u), is included in this test.
+            (
+                7e-3
+                + (0.8 - 7e-3) * (1 - 0.8) / (11.11 * pp.GIGA) * 2
+                - (0.8 - 7e-3) * 1e-5 * 3
+            ),
+            2,  # Matrix porosity is only defined in Nd
+        ),
+        (
+            models.MomentumBalance,
+            "bulk_modulus",
+            # \lambda + 2/3 \mu
+            11.11 * pp.GIGA + 2 / 3 * 16.67 * pp.GIGA,
+            None,
+        ),
+        (
+            models.MomentumBalance,
+            "shear_modulus",
+            # \mu
+            16.67 * pp.GIGA,
+            None,
+        ),
+        (
+            models.MomentumBalance,
+            "youngs_modulus",
+            # \mu (3 \lambda + 2 \mu) / (\lambda + \mu)
+            16.67 * (3 * 11.11 + 2 * 16.67) / (11.11 + 16.67) * pp.GIGA,
+            None,
+        ),
+        (
+            models.MomentumBalance,
+            "elastic_normal_fracture_deformation",
+            # (-normal_traction) * maximum_closure /
+            #    (normal_stiffness * maximum_closure + (-normal_traction))
+            (1 * 1e-4) / (1529 * 1e-4 + 1),
+            1,
+        ),
+        (
+            # Tets permeability for the matrix domain. Should give the matrix
+            # permeability
+            models._add_mixin(c_l.CubicLawPermeability, models.MassBalance),
+            "permeability",
+            1e-20,
+            2,
+        ),
+        (
+            # Test the permeability for a fracture domain. This should be computed
+            # by the cubic law (i.e., aperture squared by 12, an aditional aperture
+            # scaling to get the transmissivity is taken care of elsewhere).
+            models._add_mixin(c_l.CubicLawPermeability, models.MassBalance),
+            "permeability",
+            0.01**2 / 12,
+            1,
+        ),
+        (
+            # Test the permeability for an intersection. The reasoning is the same as
+            # for the 1-d domain.
+            models._add_mixin(c_l.CubicLawPermeability, models.MassBalance),
+            "permeability",
+            0.01**2 / 12,
+            0,
+        ),
+    ],
+)
+def test_evaluated_values(
+    model: Type[Poromechanics]
+    | Type[Thermoporomechanics]
+    | Type[MassAndEnergyBalance]
+    | Type[MomentumBalance]
+    | Type[_],
+    method_name: Literal[
+        "fluid_density",
+        "thermal_conductivity",
+        "solid_enthalpy",
+        "fluid_enthalpy",
+        "matrix_porosity",
+        "bulk_modulus",
+        "shear_modulus",
+        "youngs_modulus",
+        "elastic_normal_fracture_deformation",
+        "permeability",
+    ],
+    expected: Any | float | Literal[2370, 12540],
+    dimension: Literal[2, 1, 0] | None,
+):
+    """Test that the value of the parsed operator is as expected.
+
+    Parameters:
+        model: Model class to be tested. Likely composed of several mixins.
+        method_name: The specific method to be tested.
+        expected: The expected value from the evaluation.
+        dimension: Dimension on which the method will be invoked. If None, all
+            dimensions in the MixedDimensionalGrid will be considered.
+
+    """
+    # The thermoporoelastic model covers most constitutive laws, so we use it for the
+    # test.
+    # Assign non-trivial values to the parameters to avoid masking errors.
+    solid = pp.SolidConstants(models.granite_values)
+    fluid = pp.FluidConstants(models.water_values)
+    params = {
+        "material_constants": {"solid": solid, "fluid": fluid},
+        "fracture_indices": [0, 1],
+    }
+
+    setup = model(params)
+    setup.prepare_simulation()
+
+    # Set variable values different from default zeros in iterate.
+    # This yields non-zero perturbations.
+    if hasattr(setup, "pressure_variable"):
+        setup.equation_system.set_variable_values(
+            2 * np.ones(setup.mdg.num_subdomain_cells()),
+            [setup.pressure_variable],
+            iterate_index=0,
+        )
+    if hasattr(setup, "temperature_variable"):
+        setup.equation_system.set_variable_values(
+            3 * np.ones(setup.mdg.num_subdomain_cells()),
+            [setup.temperature_variable],
+            iterate_index=0,
+        )
+    method = getattr(setup, method_name)
+
+    # Call the method with the domain as argument. An error here will indicate that
+    # something is wrong with the way the method combines terms and factors (e.g., grids
+    # parameters, variables, other methods) to form an Ad operator object.
+    # TODO: At some point, we should also consider interface laws.
+    op = method(setup.mdg.subdomains(dim=dimension))
+    if isinstance(op, np.ndarray):
+        # This will happen if the return type of method is a pp.ad.DenseArray. EK is not
+        # sure if we still have such methods among the constitutive laws, but the test
+        # is correct, so keeping this should not do any harm.
+        assert np.allclose(op, expected)
+        return
+
+    # The method should return an Ad object.
+    assert isinstance(op, pp.ad.Operator)
+
+    # Carry out discretization.
+    op.discretize(setup.mdg)
+    # Evaluate the discretized operator. An error here would typically be caused by the
+    # method combining terms and factors of the wrong size etc. This could be a problem
+    # with the constitutive law, or it could signify that something has changed in the
+    # Ad machinery which makes the evaluation of the operator fail.
+    val = op.evaluate(setup.equation_system)
+    if isinstance(val, pp.ad.AdArray):
+        val = val.val
+    # Strict tolerance. We know analytical expected values, and some of the
+    # perturbations are small relative to
+    assert np.allclose(val, expected, rtol=1e-8, atol=1e-10)
+
+
+@pytest.mark.parametrize(
+    "geometry, domain_dimension, expected",
+    [
+        (models.OrthogonalFractures3d, 1, [0.02, 0.02**2, 0.02]),
+        (models.OrthogonalFractures3d, 0, [0.02, 0.02**3, 0.02**2]),
+        (models.RectangularDomainThreeFractures, 0, [0.02, 0.02**2, 0.02]),
+        (models.RectangularDomainThreeFractures, 2, [1, 1, 42]),
+    ],
+)
+def test_dimension_reduction_values(
+    geometry: pp.ModelGeometry, domain_dimension: int, expected: list[float]
+):
+    """Test that the value of the parsed operator is as expected.
+
+    Parameters:
+        geometry: The model geometry to use for the test.
+        domain_dimension: The dimension of the domains to test.
+        expected: The expected values of the:
+            - subdomain aperture
+            - subdomain specific volume
+            - interface specific volume
+
+    EK: This test could possibly be merged into the above test_evaluated_values(), but
+    the risk of obfuscating the implementation. It was therefore decided to keep this
+    method separate, at least for now.
+
+    """
+    # Assign non-trivial values to the parameters to avoid masking errors.
+    solid = pp.SolidConstants({"residual_aperture": 0.02})
+    params = {"material_constants": {"solid": solid}, "num_fracs": 3}
+    if geometry is models.RectangularDomainThreeFractures:
+        params["fracture_indices"] = [0, 1]
+
+    class Model(geometry, pp.constitutive_laws.DimensionReduction, models.NoPhysics):
+        pass
+
+    setup = Model(params)
+    setup.prepare_simulation()
+
+    subdomains = setup.mdg.subdomains(dim=domain_dimension)
+    interfaces = setup.mdg.interfaces(dim=domain_dimension)
+    # Check aperture and specific volume values
+    aperture = setup.aperture(subdomains).evaluate(setup.equation_system)
+    if isinstance(aperture, pp.ad.AdArray):
+        aperture = aperture.val
+    assert np.allclose(aperture.data, expected[0])
+    for grids, expected_value in zip([subdomains, interfaces], expected[1:]):
+        specific_volume = setup.specific_volume(grids).evaluate(setup.equation_system)
+        if isinstance(specific_volume, pp.ad.AdArray):
+            specific_volume = specific_volume.val
+        assert np.allclose(specific_volume.data, expected_value)
